refactor: move file actions to separate file
